api/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
elasticsearch
pydantic
aiohttp
pydantic-settings
pytest
pytest-asyncio
pytest-cov
httpx
mysql-connector-python
